only adding Buffers belonging to BufferedDataTable to MemoryObjectTracker. Isolated DataContainer (for instance kept as internals in HierarchicalClusterNodeModel) are not memory tracked as they don't get cleanly unregistered when the workflow is disposed.
We ran into a "problem" where such a table from a discarded workflow should have been swapped to disc but the node associated with the table (via current NodeContext) didn't exist anymore (neither the temp dir for the workflow). The FileUtil#createTempFile uses the /tmp folder then but it's still not clean. Now: we just register what we also cleanly unregister