Issue AP-5785: Workflow Coach: Recommending next best successor based on community usage statistics https://knime-com.atlassian.net/browse/AP-5785
Multiple occurring node recommendations must be aggregated differently for source nodes. For source nodes the frequencies are just summed up (instead of taking the average as done with all other nodes). The effect: File Reader, CSV Reader, XLS Reader are at the top of the list of the source nodes as one would expect.