Issue AP-5785: Workflow Coach: Recommending next best successor based on community usage statistics https://knime-com.atlassian.net/browse/AP-5785
Don't show the 'no selection' message anymore. Either all source nodes are shown (if no node is selected) are the recommendations for a selected node. If the user selects something in another view (except the WorkflowEditor) the recommendation list will remain as it is (since the selection within the WorkflowEditor essentially doesn't change then).