AP-6939: Continue loading settings if fails for one field
Also removes old custom class loader which was previously used to load
the java type in the field settings. Now the java type is derived from
the source type of the converter factory and only loaded when settings
from older java snippet need to be loaded.

AP-6939: Java Snippet node does not fail, if a ConverterFactory could not be
found
Signed-off-by: Jonathan Hale <jonathan.hale@knime.com>