AP-7059: SandboxedNode copy data from workflow dir
The SandboxedNodeCreator now copies the data directory from the workflow
as this is a common place to store data for the workflow which could be
needed in the sandboxed node.

As the SandboxedNodeCreator already copies the data and drop foder of the
sandboxed node and every node inside the sandboxed node the code could
easily be reused to copy the data folder of the workflow.