AP-7057: FileStoreCells in loops and remote execution
With remote execution file stores get initialized in
#mimicRemotePreExecute().
But this method is called on the nodes of a wfm in a random order
(actually
the order of creation). For file stores in loops the loop start node
needs to
create a special FileStoreHandler which is used by any node inside the
loop.
But when we try to initialize file stores on a node inside the loop
before
the loop start this special FileStoreHandler isn't created yet.

The solution is to call #mimicRemotePreExecute in the right order.
Additionally it propagates flow variables through the workflow and makes
sure the current node has the newest FlowLoopContext which is needed to
initialize the FileStoreHandler in loops.