AP-7468: NodeExecutionResult save flow variables
Added a new method to NodeContentPersistor to get the saved flow
variables and added a field to NodeExecutionResult to store them.

Also made sure the flow variables are set in the execution result when
Node#createNodeExecutionResult is called and loaded when
Node#loadDataAndInternals is called.

AP-7468 (NodeExecutionResult doesn't save flow variables pushed by the
node)