Don't complain about missing extension point of port class is an interface
You can define a port object interface which is implemented by several
classes. In the code you use the port type of the interface but without
this change the registry complains that the interface is not registered
at the extension point. However, you cannot register it, because there
are usually not (de)serializers for the interface which are required.