AP-7792: Ensure that all possible values of the target col are included
Before the leaf distributions contained only those classes that actually
appear in the input table. This causes problems because the modelSpec
stores the tableSpec of the input table which can contain possible
values that do not actually appear in the table.

AP-7792 (Soft voting feature in the tree ensemble works only if the
target variable has only 2 possible values)