AP-7827: fixed 'setDirty' for SubnodeContainers
The setDirty() wasn't properly propagated to the parent workflow
for any node that is inside a wrapped metanode (subnode). So far this
hasn't been an issue as all actions that affect the subnode have been
done inside the subnode and the workflow editor(!) was flagging the
outer workflow dirty.

AP-7827 (New usage configuration dialog for wrapped metanodes)