Moving 'old' features back into desktop build
Previously tried to have all non-core extension part (e.g. json)
be seperately installed to allow them to be uninstalled.
However, that breaks the update (features will not be available
when going 3.4 -> 3.5) and hence they are again part of the
standard build and only new bundles will be separately installed