AP-8680: Updated enable status of columns in table selector
All the options in the Table creator right-click menu are relative to
the currently selected column(s). This commits ensures they are only
active when at least one column is selected, previously this was
intended but did not work consistently.

AP-8680: Table Creator: Can't change column type if the column is not selected