AP-9228: reset filechooser on cancellation, update history
* added committing of editable JComboBoxes to
commitComponentsRecursively to ensure that the selected item is the item
displayed in the textfield of the combobox
* added call of updateHistory in DialogComponentFileChooser to ensure
that the last committed values are shown in the history when re-opening
the dialog
* removed condition in ReadTableNodeModel that saves settings only when
the chosen file name is not null. This caused that the entered file name
is not reset when its key is unknown to the settings model.

AP-9228 (File chooser component does not reset after 'cancel' in dialog
is clicked)