AP-6886: reset virtual ParChunk metanode before deletion
I think this is somewhat a workaround exploiting the fact that during
#reset we do not check the downstream nodes of a WMN to be in execution.
This is different to when we delete an executed node in a WMN ... then
we do check if there are downstream nodes in execution.

In any case this fix doesn't make things worse than before.

AP-6886 (Parallel chunk loop fails if in wrapped meta node)