AP-9524: Move "createTokenizerSettings into CellSpliterUserSettings
Converting it to a member to remove the null check, adding CheckUtils
checks for functions where this was used instead (which are never called
with null, but theoretically, it's possible, so why not).

Will need this function from other places (e.g. node model or dialog)
soon.

AP-9524 (Cell Splitter: "Split header column")

Signed-off-by: Jonathan Hale <jonathan.hale@knime.com>