AP-10188: Fix NPE for layout editor on windows
The visual layout editor control was being created before the advanced
tab control. For linux and mac this wasn't an issue because the JSON
document was already initialized, but that was not the case on Windows
since Windows uses a text area. Initializing the visual layout editor
control after the advanced control fixes this issue.

AP-10188 (Integrate visual layout editor into layout dialog)