AP-9255: Set min size on basic scroll panel and update layout
Setting the min size on the scroll panel will ensure that scroll bars
appear when the window is smaller than the minimum size. However, in
order to change the minimum size of the scroll panel after it has been
defined the layout for this must be re-positioned by requestLayout().

AP-9255 (Wrapped Metanode - Node Usage Configuration Tab is missing
Dialog Pane)