AP-13325: LOGGER.assertLog's removed to assert presence of NodeContext
In order to determine whether nodes are streamable (for the node
repository) some node models need to be instantiated (in case pure
reflection fails). Those node models should be disposed via
NodeModel#onDispose afterwards. However, this method expects a
NodeContext to be present but we don't have one (because the
instantiated node models are not part of a workflow, i.e.
NodeContainer). Thus we remove the LOGGER.assertLog's (triggered if no
node context is given) and catch the respective error at respective
places (something that actually should never happen since a node context
is not available on node instantiation either, afaiks - only during
configure and execute).

AP-13325 (Streamable source nodes are not correctly flagged as such
anymore in node repo)