AP-13503: Prevent fvm change event on tree update
If a flow variable was selected for a FlowVariableModel four change
events were called once the tree element dialog tree element with the
flow variable was drawn. These events did not really reflect any change.
This was the case because the combobox model was updated by removing all
elements and adding the available elements. Therefore the selecting
changed because the selected element was not available anymore. This was
solved by using a new DefaultComboBoxModel which is only used after all
element were added.