AP-13847: Release resources held by Buffer
Resources include TableStoreCloseableRowIterators and registered
MemoryAlertListeners. When such resources are owned by the Buffer, they
are released on clear. When such resources are owned by a
CloseableRowIterator iterating over the Buffer, they are released when
that iterator is closed or garbage-collected or when the Buffer is
cleared.

AP-13847 (Table does not unregister memory alert listeners on clear)