AP-11595: flow var tab not filling width
There isn't a concrete reason to do it this way but this is as in the
last PR that Loki created. During code review I "temporarily" set the
flag to true but didn't revert it before rebase/merge.

The only reason to keep this false is that the code will be less complex
(could also be simplified if we knew it remains false).

AP-11595 (Improve formatting of long flow variable names in flow
variable tab)
https://knime-com.atlassian.net/browse/AP-11595