Experimental Spliterator Implementation.
Recursively splits the rows of a BufferedDataTable in half.
Like an Iterator, delivers the DataRows belonging to the Spliterator's partition.
See org.knime.ext.benchmarks.tests.benchmark.buffer.BufferedDataTableSpliteratorBenchmark; for an example on how to use the Spliterator to process a table in parallel. I increased the filestore visibility to use it in the Benchmark.