AP-13980: Fix loading for overwritten array configs in legacy workflows
Prior to 4.1.0, array configs were fully represented in the flow
variable tree and therefore could be overwritten using flow variables.
With 4.1.0, we introduced array flow variables and in 4.2.0 we
introduced an extension point for flow variables that allows to define
even more complex types. As a consequence those complex flow variables
are now displayed as leafs in the flow variable tree. However, if an old
workflow overwrites an array config with a flow variable, this array
config still needs to be displayed in the flow variable tree as it used
to be.

AP-13980 (Implement DialogComponentFileSystem and FSLocationFactory)