AP-13907: Special handling of an edge case
The following condition will cause the workflow executor to fail on
purpose:
Workflow executor executes a workflow that contains a capture scope that
contains a metanode that contains a source node with no input connections.
Otherwise it will fail later, e.g. in the Workflow Writer, because the
mentioned source node won't be replaced by a reference reader node
(because the source node is not part of the virtual scope).

AP-13907 (Inner-workflow execution of 'workflow-producing' workflows results in broken 'static-input' references)