AP-14534:Fixed merge join columns.
JoinSpecification#columnIncludeIndices was illegally used in rowProject.
Renamed rowProject to rowProjectOuter to make clear that this is only for unmatched rows from left/right table (when separately output).
Renamed columnIncludeIndices to getMatchTableIncludeIndices to make clear that this is only for matches or getSingleTable().

AP-14534 (Joiner merge join columns is broken)