AP-13067: Make component nodes not part of scope
Make isMemberOfScope return false for nodes inside a Component. They
have a FlowScopeContext on the stack which is what was checked to
determine if the node is a member of a scope. Now we check that this
FlowScopeContext is not a FlowSubnodeScopeContext.
This is done because isMemberOfScope is used to determine if a "Select
Loop" option is shown in the context menu. This option should not be
shown for nodes inside a Component.

AP-13067 ('Select Loop' shown in context menu of component input node)