AP-14915: Special case handling of metanode successors
I.e. if a metanode is directly connected to its parent it needs a
special handling to determine whether there are successors in progress.

Second iteration for that fix: now we also taking into account that the
parent can be in 'executing' state.

AP-14915 (Execution of one metanode blocks editing of other metanodes (both contained in super metanode))