AP-15542: Ignore flow variables overwriting settings for inactive nodes
Because even though the node is 'executed' (i.e. skipped) the flow
variable overwritting a setting might not be present. And instead of
ignoring only those it is clearer to ignore all in case of inactive
nodes (the node settings for inactive nodes will be added to the
workflow summary nevertheless, but without considering overwriting
flow variable values).

AP-15542 (Workflow summary generation fails for nodes in an inactive branch with absent flow variable)