AP-14408: Add ClosableShuffler and deprecated Shuffler
The Shuffler creates an intermediate table which can not be cleared
after the caller is done with the shuffled table because the caller has
no reference to it. This can cause an additional temporary file to be
created on each shuffle (which cannot be cleared).

The new ClosableShuffler has a #close method which allows to clear this
intermediate table once the caller is done with the shuffled table.

AP-14408 (DL Keras Learner creates many temporary files which are not
deleted during execution)