AP-15550: Add context for template workflow root
During updating and sharing of metanodes, sometimes `configure` and
`onDispose` methods of the contained nodes are called when they are
still attached to the template root workflow manager. This causes issues
if the nodes try to access the context in these methods. This commit
provides a dummy context for the template root wfm.

AP-15550 (NullPointerException in Call Workflow (Table based) node)