AP-15557: Minimal fix for the NPE
Which let's the Parallel Chunk Loop execute but doesn't solve the
underlying problem. I.e. captured workflows that reference data in the
temporary metanode created by the Parallel Chunk Loop can't be properly
used outside of the Prallel Chunk Loop (because the data is gone then -
it would need to be stored in the Parallel Chunk Loop End/Start,
mitigated by the PortObjectRepository).

AP-15557 (NPE when capturing workflow which is part of a Parallel Chunk Loop)