AP-16036: Match ordering of configuration nodes in component description
The component layout editor provides means to define an ordering on
contained configuration nodes. Previously, this ordering would not be
respected in the node description of the component. This commit fixes
that issue. Configuration nodes which are newly added will appear at the
end of the ordering in the order in which they were added.

AP-16036 (Component description is not matching new reordering of configuration nodes)