AP-13893: Allow flow variables to be initialized without password
... if a PortObjectReferenceReader is loaded with credentials-flow
variables which don't have a respective entry in the credentials store
(or there is no store set) of the respective workflow.
Allows one to create (i.e. load) PortObjectReferenceReader nodes with
credentials-flow variables within a workflow with an empty credentials
store (e.g. workflow fragments for ID).
In that case no password is set, which is desired for a workflow fragment
stored with a workflow port object.

AP-13893 (Problem with credentials flow variables when injected via reference reader node)