AP-15879: Get virtual context from node connected to metanode outport
Because the metanode itself has no scope context set, but the contained
nodes. If a metanode port is part of a virtual scope context, the output
data is stored with the node responsible for the virtual scope (e.g. the
Workflow Executor)

AP-15879 (Fix virtual scope context retrieval from metanode)