AP-15882: Recursive retrieval of flow virtual scope context revised (again)
Because with the previous commit I ignored the fact that components
don't know their parent flow context.
That's what we do now:
If the parent container is a component, check the context of it.
If the parent container is a workflow, follow all workflow incoming
connections and check the connected nodes.
Stop once we are at the root level.
Another change: we only try to find the virtual scope context for the
capture start node (and not for the 'static' input anymore).

AP-15882 (Problem with connections reaching into a capture scope whose
source is the parent workflow (i.e. metanode port))