AP-15078: Add toString representation to WorkflowPortObjects
This has the effect that WorkflowPortObjects will be represented by
their custom name when converted to a cell with the "Model to Cell"
node. If no custom name is set, the default workflow name is used.

AP-15078 (Integrated Deployment: String representation of a workflow should be its custom workflow name if set)