AP-15078: Make WorkflowPortObject#getSummary return custom workflow name
This has the effect that WorkflowPortObjects will be represented by
their custom name in the output table view when converted to a cell
e.g. with the "Model to Cell" node. If no custom name is set, the
default workflow name is used.

AP-15078 (Integrated Deployment: String representation of a workflow should be its custom workflow name if set)