AP-17289: Expose BDT.getDelegate via Node
This is needed by knime-python to obtain the (Arrow) batch store of a
table stored using the columnar backend, which is a prerequisite for
efficiently sharing data tables with external tools such as Python.

Note that we do not make BDT.getDelegate directly public since the
class is fairly visible to node implementations.

AP-17289 (Implement putDataTable for real)