AP-17192: Reimplement FileStoresInLoopCache with HashSet to allow out-of-order insertion
When the columnar backend serializes FileStore cells asynchronously,
it can happen that the insertion order is not the same as when writing
synchronously. As we do not expect the set of FileStoreKeys to be too
huge, we keep them in memory and just use a HashSet. We make sure
we return an ordered table of FileStoreKeys when the cache is closed.

AP-17192 (Columnar backend memory issues)