AP-11762: Untangle Workflow and Metanode in the workflow schema.
A workflow is no longer a node. Instead, a MetaNode entity is introduced with a reference to a Workflow,
similar to how Component works.
Started replacing the *Persistor interfaces with cleaner *LoaderInterfaces that only contain the logic/process parts of the Persistor interfaces.

AP-11762 (More options to save workflows (folder, knwf, mongo, ...))