AP-15779: Track ContainerTables returned by the TableBackend
Fixes the following scenario that otherwise causes memory leaks when fast tables are used:
Some nodes (e.g. GroupBy) create temporary tables that are not part of the output and also not
referenced by any of the output tables. Such tables created via a DataContainer are tracked in the local
table repository, and are cleared when the node is reset. However, this was not the case for VirtualTableExtensionTables
which are also ContainerTables (because they are a ExtensionTable) and although they don't hold data of their own, iterating
over them opens resources that require closing.

AP-15779 (More Memory Leak Hunting)