AP-17490: Add CollectionValueFactory#getElementValueFactory
Needed to extract the DataType for dealing with CollectionValueFactories
in a generic way. Since we also initialize them with an element
ValueFactory, it is reasonable to also provide a getter.
The method is used to determine the DataType corresponding to the
ValueFactory as well as for creating the LogicalTypeTraits (i.e.
serialization).

AP-17490 (VirtualTableExtensionTable and DataCellValueFactory don't play
along)