AP-17501: Fix handling of CellCollections in DictEncodedDataCellValueFactory
CellCollections serialize multiple DataCells while writing the outer
DataCell to the DataOutput, this was overlooked in the first iteration
of this feature. Now we construct a string containing all DataCell class
names separated by semicolon and store that alongside the binary
contents of the DataCells so that we can properly deserialize them.

AP-17501 (Let DataCellValueFactory store serializer alongside data in
dict encoded struct)