AP-17525: Introduce PortTypeClassMapper extension point
This extension point allows to add a PortTypeClassMapping that maps
between PortObject classes. The extension point is not meant for public
use because in order for it to work, the new class must fully support
the old class during loading (otherwise components that used the old
class won't load anymore).

AP-17525 (WEKA models should not be kept in memory)