AP-17841: Type AbstractDataCellWriteValue on DataValue
The read values we return are actually proxy objects that implement
DataValue but not DataCell (because the former is an interface while the
latter is not). They can therefore be cast to ReadValue which in turn
can provide a DataCell via ReadValue#getDataCell().

AP-17841 (Class Cast exception when materializing Table for Python3)