AP-17832: Use IntKeys for DictEncoded DataCell serialization
Because we do not have a table-wide dict element cache yet,
we do assign new keys for the dict encoded values in each batch.
If we just use byte keys, this limits us to 255 batchs in the best
case, which is too few. Int should be enough though.

AP-17832 (Dict Encoding runs out of byte keys)