AP-18074: Provide NodeContext when accessing data in the table view
AdapterCells need the TableBackend in order to work properly with
BlobDataCells,
and they use the NodeContext to retrieve the TableBackend for the
current workflow.
However, in the view the NodeContext was not available which meant that
the configured
default backend was used.

AP-18074 (Sorting tables containing DictEncoded DataCells fails in view)