AP-18248: fixed behavior of WorkflowManager#canPerformReset
As per comment in ticket on the previous behavior (which is now fixed):
Clearly it’s order-dependent which value is returned. If enumerating the nodes
first returns an executing node (line 8) it will return false.  If it first
encounters an executed node (line 11) the result will be true. This is not
deterministic and we should change the logic to first check if there are any
executing nodes (reset not allowed/ false) and only then check if the reset
makes sense (= contains executed nodes).

AP-18248 ("Reset" workflow action resets nodes with executing successors)