AP-18593: fixed blob reference in cell-by-cell copy
This is the actual fix of the problem. In case a NON-BufferedDataTable
is copied into a BufferedDataTable the blobs need to be copied in order
to "register" the blobs with the workflow. This part did work well also
before the fix and the blob file was properly created (just right above
the modified line). However, the original blob cell (BlobDataCell) was
retained and passed into the wrapper, which is wrong. This led to
wrapper cells representing blobs outside the workfow. There was (and is)
and assert statement in the code to check that the blob addresses are
the same, and this assertion fired. (Error message of the assertion
improved also.)

AP-18593: Scrambled Output in Blob Columns With Old Table Backend