UIEXT-200: Don't keep HiListeListeners as weak map-values in SelectionEventSource
This is not required at the moment (premature optimization) because
HiLiteListeners (and HiLiteHandlers) are currently always removed when a
single node view or composite view is being closed.
Needs to be re-visisted as soon as node views are embedded in AP.next,
see NXT-818.

UIEXT-200 (HiLiteListeners sometimes don't get removed when composite view is closed)