Fix ValueFactory type returned for lists of DataCells
We use a fallbackFactoryProvider to use the default DataCell serialization
if no ValueFactory is available for a DataType. But if we have a list
of DataCells without ValueFactory, we still used the fallbackFactoryProvider
for List<DataCell> instead of just DataCell. This lead to nesting types in lists
one level deeper than needed.