UIEXT-277: SelectionTranslationService and Selection refactored
Refactored to translate from a set of row keys to a list of string and
back.
No more SelectionTranslationService based on InputStreams
(string-based only now).
SelectionEvent returns a list of strings (called selection) which
doesn't need to be a list of row keys (up to the node implementer). If
row key to string list translation fails, it will contain an error
instead.

UIEXT-277 (Generalize Selection Events)