UIEXT-227: 'canExecute'-property removed again from NodeViewEnt
Because the frontend now bases the 'canExecute'-decision exclusively on
the node state (!idle -> canExecute=true).
Reason: There are situations where an idle node could be executed.
However, idle also means that no input spec is available to the node and
we currently don't update the input spec on execution. I.e. the node
dialog very likely doesn't reflect that then. That's why we just don't
allow execution in those cases for now.

UIEXT-227 (Nodes should be able to open dialog even if no data is connected)