UIEXT-227: 'canExecute' to NodeInfoEnt added
Which indicates when a node can be executed from within a node view.
Decision is based on the node state and available input specs - that's
why it needs to be determined on the backend side.
Test added, too.

UIEXT-227 (Nodes should be able to open dialog even if no data is connected)