AP-18412: Add DenseBitVectorValueFactory to add FastTables support
In this first attempt, BitVectors are stored as hex string. While
this sounded promising on the first try, we noticed that the hex
string representation of a DenseBitVector is truncated after a user
defined length, meaning we might not store the full value of the BitVector.

Will be improved in AP-19224

AP-18412 (Implement RDKit Extension Types)