AP-19515: Fix adapter WriteValue setting from ReadValue
For some reason the call chain inside a component is different from the call
chain top level in the workflow, which made this problem surface.

When an AdapterCellWriteValue is set with a ReadValue, we immediately copy the
values of the accesses instead of deserializing and serializing the data cell.
This saves us a roundtrip and prevents a ClassCastException because our
AbstractAdapterCellReadValue did not implement AdapterValue. It now does,
for the sake of completeness.

AP-19515 (Reading AdapterCells from outside of knime-core fails in components when using FastTables)