AP-3276: Add alphanumeric string comparator; refactor row comparator
Add comparator to compare string-compatible values alphanumerically
instead of lexicographically. This order is also called
"natural sort order", but it seems that "alphanumeric" is the term
used in programming.

Refactor the RowComparator to use the builder pattern in order to avoid using `-1` publicly to mean "compare row id".

If, in the future, the same column should be comparable
multiple times, the underlying row comparator implementation can be adapted.

AP-3276 (Sorter node: Support for Natural Sort Order)