AP-19761: lock workflow when loading content inside component
Fixes a regression which, I think, is only due to premature status checks in
the loaded workflow. Previously, the workflow would be locked only inside the
'loadContent' call. When the lock is released (also within that call), the
WorkflowLock is set to "computeNewState", which includes fetching error
messages from the contained nodes. However, that is too early since not all
innner content of the component is loaded, e.g. the "component input node".
This would then cause the "getNodeErrorMessage" to be called, which changed in
4.6. (I believe the change made in 4.6 is good and safe but there was workflow
locking missing.).

The locking could also be done further outside but looking at other similar
calls it seems we/I wrap only that single method invocation.

AP-19761 (Importing and loading (complex) workflows previously exported without data might fail)