UIEXT-654: 'hasNodeView'-predicate introduced
Which indicates whether a dialog's node also has a node view. It
is used within a 'isModelSettingAndHasNodeView'-utility function which
in turn controls the visibility of the re-execution icon and whether
settings are marked as dirty (both of it is only relevant if it's a
model setting _and_ the node has a node view to update).

UIEXT-654 (Remove 're-execute' icon from settings in dialogs without view-preview)