AP-19803: Catch errors when fetching rows in the table view
If the underlying data is broken (because of a bug somewhere else)
getting the next row can fail with an exception. This was not handled
nicely. We now catch this exception, log it to the console, and show a
dummy row that indicates that an error has happened.

AP-19803 (Memory leaked when Python node is reset / filled with weird geospatial data)