AP-19933: Recursively check for 'Reference Reader'-nodes
... when reading in a workflow with the Workflow Reader.
Because the data of all Reference Reader-nodes (also the ones that are
nested in metanodes) needs to be copied into the Workflow Reader.
Reference Reader-nodes are wrapped into a metanode if they represent
different ports of the _same_ original node.
(components can be ignored, because they are never collapsed into a
compnent)

AP-19933 (Data of Reference Reader-nodes not copied over to Workflow Reader when nested in metanode)