AP-19698: Only merge and keep bounds if the current type is bounded
In previous versions this was implicitly the case because merging a
bounded with an unbounded value would set the bounds to null because the
unbounded value didn't have bounds.
However, that behavior was incorrect because merging a column of a
bounded type that has no bounds (e.g. because all values in the column
were missing) with another column of a bounded type that has bounds
(e.g. int with int) should take on the bounds of the second column.