AP-21449: Buffer log messages before instance location has been set
This commit adds logic to buffer messages that are logged before the logging is configured.

The logging is initialized if an instance location is already set during initialization of the `NodeLogger` class or via the service framework, listening for the event that the instance location has been set.

To make buffered log messages show up with the correct timestamp, we inject the instant at log time into the log event being replayed from the buffer.

AP-21449 (Make Node Logger safe to use during application startup)