UIEXT-1534: Adjust the default transformation from cell size to rendered size
With that change, if the renderer does not overwrite this method, the
rendered data will be requested in the size of the cell (i.e. scaled)
and not in the preferred size.

The preferred size was originally used here to enable requesting knip
thumbnail images in the correct aspect ratio (to scale these down in the
TableView)

UIEXT-1534 (RDKit cell content resizing is not working)