AP-21327: SubWFM to send change events to SubnodeContainer synchronously
... which will help fixing a number of test cases. The problem, so far,
is only
visible in test cases: A change to the WFM inside a SubnodeContainer is
made,
which will queue a change event (e.g. node removed, ports changed). So
far this
was done always asynchronously. This caused resets on the workflow (see
SubnodeContainer's listener implementation), which would come after
other
changes in the test. So far I believe it's only affecting test cases. It
sounds
wrong in either case so this change should fix the root cause.

AP-21327 (Fix test case EnhAP20402 SubnodeEndWaitingForNodes)