UIEXT-1552: Make `RowContainer#close` throw specialised `IOException` instead of general `Exception`
Otherwise, all consumers must annotate the general Exception as well, which makes using the v2 API slightly annoying.

`#finish` already throws an IOException, so this does not add a new exception type to the class.

This is a breaking change, but the API is marked as experimental (`@noreference`).

UIEXT-1552 (Web UI for Row Filter Iteration 1)