AP-22005: Fix null handling in level mapping
Before this change, we would map `null` to Level.ALL, which would be wrong for an upper bound, where it should have been mapped to Level.OFF (the highest upper bound).

Instead of conditionally using the fallback, we can also just return `null` meaning the bound is unspecified, since Log4j filters can support a null level (exactly meaning unbounded, which is also the semantics when omitting the minimum/maximum element in the config).

This means, that if a `null` level is passed, we return a `null` mapped level. Then in whichever context the original value originated, minimum or maximum, the mapped `null` value means the same.

Why do we want to map to `null`?
When we had configured an appender with the following XML config:
- minimumLevel=OFF
- maximumLevel element omitted entirely
Log4j would interpret the config as the range (minimumLevel=`Level.OFF`, maximumLevel=`null`).
Our mapping would result in (minimumLevel=`LEVEL.ALL`, maximumLevel=`LEVEL.ALL`), since `Level.OFF` was mapped to `Level.ALL` through a default case and `null` to `Level.ALL`.
This results in a filter that filters out any log message, since no level passes this filter (nothing can be lower than the maximum here).
However, later the KNIMECorePlugin of workbench would set the _maximum_ of logfile and stdout appenders to `LEVEL.FATAL` effectively re-enabling the filters again (since 2013!).
Now, when a configuration lowers the maximum of stdout to, e.g. WARN in order to only report ERRORs and FATALs on stderror, the plugin would raise the maximum of stdout to FATAL, making stdout and stderr appenders overlap.

A fix was applied to retain the maximum of an existing range and only fall back to FATAL in case no range was present (in order to be able to set the new minimum and keep old behavior). However, this now left the filter “broken” instead of “fixing” it (and producing duplicates on stdout and stderr if the latter was enabled).
AP-22005 (Avoid duplicate log lines on stdout & stderr)