AP-23029: special case blobs when updating domain (efficiency)
... noticed this when writing the test case for the associated ticket: blob
cell columns are always unwrapped when updating domains, just to discard the
blobs immediately afterwards as they are not considered for domain updates.
Same is true for metadata updates. Slighly changes to contract of
MetaDataCalculator.update(DataCell) but there are only two implementations of
it in KNIME.

AP-23029 (Core Data API: Make `BufferedRowContainer#setFrom` faster when RowRead and RowWrite are both "buffered")