AP-23029: made RowWrite#setFrom performant for row backend
The fixes are minimal but not ideal: This now somewhat contradicts with
the expectation that the client can call `setFrom` and then overwrite
individual columns later via `getWriteValue(i).setValue(...)`. At this
point this is not done by any client so I would accept this as a pseudo
bug fix.

Also an observation: `setFrom` in this new form retains cell types
(think of mixed columns with IntCell and DoubleCell) whereas going via
cell-by-cell copy all values are reduced to the type of the column (e.g.
DoubleCell.TYPE).
My expectation would be that the row backend retains cell types whenever
possible.

AP-23029 (Core Data API: Make `BufferedRowContainer#setFrom` faster when
RowRead and RowWrite are both "buffered")