AP-23175: Variable filtering only applied to current context
This was a longer debate, also dicussed in the ticket. The global filtering of
variables is not feasible with the current design of using single flow object
stacks with a node. As discussed with Leo, Armin and Lukas we will change the
functionality such that filtering of variables only applied to the current
scope context.

This commit partially brings back the changes that were made in
6cd4ce9981e8e57d8050f0c3f5081e5ae50b6e0d

and then again removed/changed in
5a4437f924ad20589dce60f125e955b2a0d5b165

AP-23175 (Variable Filter: Merged filtered stacks are unexpected)