AP-23380: fixed (another) mem leak - FlowObject leakage
again, this fixes should not be necessary but 3rd party extensions might
register code in static data structures (e.g. lambdas having instance
pointers
to the NodeModel), so that the nodes are not collectable. As part of the
ticket
we make a best effort to cut the connection between the 3rd party
NodeModel and
the core data structures.

AP-23380 ((Internal API:) Intended breaking point to reduce impact of
possible memory leaks in 3rd party extensions)