AP-23487: Revert "AP-23487: Add LongPathCheckerEarlyStartup"
This reverts commit d5e2423bc9691a94236f26fc334ba7fe2781a657.

Unfortunately it's not possible in Java to determine if there is a long path
issue. Java happily creates nested long path folders with files, which it can
also read. Other Windows programs, like notepad.exe, will not be able to read
it but Java is. More details in the ticket.

AP-23487 (Windows: Log error to knime.log when detecting Windows path length limit (256 chars, unless changed in Windows registry))