AP-23812: added org.knime.core.workbench (larger changes)
... this is a larger change, moving code from the old explorer code in
org.knime.workbench.core to knime-core (note the slight difference in package
names -- accepted as the old code will eventually die). This commit is not just
a move of code but a larger refactoring of the concept of mount table, mount
points and adapters -- such as in MUI and CUI -- to use entries in a mount
table.

AP-23812 (Move MountTable to new org.knime.core.workbench plug-in (retire workbench code)