AP-24224: run "setInactive" completely invisibly
The previous solution would cause deadlocks since the thread calling
'runInvisible' would own the wfm.lock but would also need to wait for
a slot in the thread pool to become available. That is then a
deadlock(-like) as all other threads from the pool could wait for
wfm.lock to become available.

AP-24224 (Invisible ThreadPool threads to not stay invisible for the
remainder of their lifetime)