AP-24217: Create and dispose workflow log appenders with workflow project lifecycle
- One workflow log appender is created and attached to the root logger when the workflow project
  is created
- The appender is removed and closed when the workflow project is closed (via the resource cache)
- Reference counting is used to decide when to finally close the appender. Needed when the same workflow is opened as
  multiple workflow projects, e.g. by Simple Streaming

Fixes bug of double-checked locking in `addWorkflowDirAppender` that could previously get an
appender that was about to be closed in the `MyWorkflowListener`, attaching that to logger instances
again, effectively "leaking" the closed (workflow log) appender.
What made matters worse (and that problem visible) was that this leaked appender would complain
about being closed even before evaluating its filters (and deciding that it was not supposed to log
the message anyway): "log4j:ERROR Attempted to append to closed appender '<appenderName>'.
The leak is visible (good), but it could produce log statements very rapidly
(multiple gigabytes in a few minutes) without a runtime means to temporarily silence the complaints.
AP-24217 (Attempts to log to closed appenders (again))

This also fixes some other bugs:

- Fix logging "withoutContext" being ignored:
    This avoids a recursive logging scenario in the NodeContext.

- Do not confuse appender levels with logger levels when appender levels are modified:
    Loggers emit log statements of a particular severity and appenders are consuming
    log statements in the range of their two severities (both inclusive).

    If we would "update" logger levels based on minimum appender levels,
    setting one appender to "DEBUG" would cause all loggers to emit also their "DEBUG" messages,
    where otherwise they would not have. So now appenders that already accept "DEBUG" get the
    debug messages from these loggers, too.

    Constraining individual loggers (e.g. noisy or unimportant ones) to some log level is a useful user configuration
    that we would otherwise change.