AP-24217: Create and dispose workflow log appenders with workflow project lifecycle
- One workflow log appender is created and attached to the root logger when the workflow project
  is created
- The appender is removed and closed when the workflow project is closed (via the resource cache)

Fixes bug of double-checked locking in `addWorkflowDirAppender` that could previously get an
appender that was about to be closed in the `MyWorkflowListener`, attaching that to logger instances
again, effectively "leaking" the closed (workflow log) appender.
What made matters worse (and that problem visible) was that this leaked appender would complain
about being closed even before evaluating its filters (and deciding that it was not supposed to log
the message anyway): "log4j:ERROR Attempted to append to closed appender '<appenderName>'.
The leak is visible (good), but it could produce log statements very rapidly
(multiple gigabytes in a few minutes) without a runtime means to temporarily silence the complaints.
AP-24217 (Attempts to log to closed appenders (again))