UIEXT-2761: Revert changes from UIEXT-2761
See UIEXT-2793

a) Revert e91bf05e9c21
b) Revert 37ca352e4187
c) Revert 470340e04357

The commit c) "fixed" a behavior of the NameFilterConfiguration in cases where a column was at the same time included and excluded.
With b) we avoided the regression that this behavior had been expected by some users adding a feature flag.
With c) we further tried to avoid that the behavior of such old nodes changed even when the user opened the dialog and saved it again.
But the changes in c) did not suffice to solve that problem, since the configuration used when saving a dialog is not the same as is used when loading the dialog.
The same holds true for modern dialogs, where we did not touch the problem at all for nodes that have a webUI node model. To avoid inconsistent behavior and partial solutions, we this decided to revert all of these changes again.
"fix" is written in quotes, since the behavior does only influence cases where either included or excluded columns are overwritten by flow variable and in this case the new behavior has the effect that the flow variable is ignored or partially ignored.
We want to (and have to) properly solve it for nodes when we migrate them to nodes operating on DefaultNodeSettings since those do only have the new behavior. For that, we create the followup UIEXT-2793.

UIEXT-2761 (ColumnFilter overwritten by flow variable behavior changed)