NXT-3407: Do not reset component project on load...
This was previously done in the routine that would reset nodes in a workflow which do not have all (non-optional) input ports connected. But this does not make sense for component projects (which are still represented by a SubNodeContainer instance), since it will always be a direct child of WorkflowManager.ROOT and thus never have input ports connected.

NXT-3407 (Input data of component project is lost when editing it)