AP-24599: Call flushToFileStore during WorkflowToolCell creation
Non-local workflows are copied to a temporary directory when they are
read. If we don't flush immediately, that directory might be cleaned up
already by the time we try to flush, which would raise an exception.
Calling flush is save because flushToFileStore does not flush again if
it has been flushed already.