AP-25290: Install prefs syncer after workspace location has been set
The prefs syncer uses Eclipse preferences. If they are used, they might initialize the workspace location to
its default value if it is not already set. This is not a problem if the application is started with `-data`.

Before a change to URL resolution logic, the WorkbenchActivator was implicitly called after the profiles had been loaded
and thus after the instance location (workspace) had been set. With the change, class loading of
`org.knime.core.internal.knimeurl.URIToFileResolveImpl` could happen before the workspace had been set which implicitly
started the `WorkbenchActivator` bundle, which installed the mountpoint prefs syncer.

Installing the prefs syncer right before initializing the mount table but in the `AFTER_PROFILES_SET` early startup
phase, fixes the problem.

AP-25290 (Nightlies: WorkbenchActivator initializes workspace location to default location)