AP-25368: Make WorkflowManager.cleanup a bit more robust
It otherwise throws an IllegalArgumentException in the following
scenario:
WFM.cleanup iterates it's contained nodes and calls 'cleanup' on them,
too. However, there are some nodes (e.g. WorkflowExecutor) that remove
nodes on clean-up, too (it removes the auxiliary metanode next to it
which is used to execute the workflow segment). Hence, the clean-up
logic try to call cleanup on a node that doesn't exist anymore.

AP-25368 (IllegalArgumentException when closing a workflow with an auxiliary metanode for workflow segment execution (e.g. Workflow Executor))